Rename Parser file Resolve minor merge issues
