+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -8,17 +8,16 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.logic.commands.AddCommand;
+import seedu.address.logic.commands.AddClientCommand;
 import seedu.address.logic.commands.AddRentalCommand;
 import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.EditCommand;
+import seedu.address.logic.commands.EditClientCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.EditClientCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
@@ -56,13 +55,13 @@ public class AddressBookParser {
         switch (commandWord) {
 
         case AddClientCommand.COMMAND_WORD:
-            return new AddCommandParser().parse(arguments);
+            return new AddClientCommandParser().parse(arguments);
 
         case AddRentalCommand.COMMAND_WORD:
             return new AddRentalCommandParser().parse(arguments);
 
         case EditClientCommand.COMMAND_WORD:
-            return new EditCommandParser().parse(arguments);
+            return new EditClientCommandParser().parse(arguments);
 
         case DeleteCommand.COMMAND_WORD:
             return new DeleteCommandParser().parse(arguments);
