+++ src/test/java/seedu/address/logic/parser/EditClientCommandParserTest.java
@@ -51,7 +51,7 @@ public class EditClientCommandParserTest {
     private static final String MESSAGE_INVALID_FORMAT =
             String.format(MESSAGE_INVALID_COMMAND_FORMAT, EditClientCommand.MESSAGE_USAGE);
 
-    private EditCommandParser parser = new EditCommandParser();
+    private EditClientCommandParser parser = new EditClientCommandParser();
 
     @Test
     public void parse_missingParts_failure() {
